requirements.txt
transformers==4.31.0
sentence-transformers==2.2.2
pinecone-client==3.1.0 
accelerate==0.21.0 
einops==0.6.1 
xformers==0.0.20 
gradio==3.48.0
replicate==0.25.1
pypdf
llama_index
python-dotenv